tst_qaudiosink: Replace max qint64 with max int
For testing audio sinks we can set up a mock source that always reports
available bytes. This used to use std::numeric_limits<qint64>::max(),
but this caused issues on 32-bit Android.

This patch replaces the available bytes with max int instead.

Pick-to: 6.10 6.9 6.8

+++ tests/auto/integration/qaudiosink/tst_qaudiosink.cpp
@@ -29,7 +29,7 @@ class AudioPullSource : public QIODevice
 {
 public:
     AudioPullSource(bool isContinuous = false)
-        : available(isContinuous ? std::numeric_limits<qint64>::max() : 0),
+        : available(isContinuous ? std::numeric_limits<int>::max() : 0),
           m_isContinuous(isContinuous)
     {
     }
